refactor(rev-detail): move action from DataController
