+++ src/Resources/config/routing/data.xml
@@ -24,7 +24,7 @@
         <default key="deleted">0</default>
     </route>
     <route id="emsco_view_revisions" path="/revisions/{type}:{ouuid}/{revisionId}/{compareId}" methods="GET">
-        <default key="_controller">EMS\CoreBundle\Controller\ContentManagement\DataController:revisionsDataAction</default>
+        <default key="_controller">EMS\CoreBundle\Controller\Revision\DetailController:detailRevision</default>
         <default key="revisionId">0</default>
         <default key="compareId">0</default>
     </route>
@@ -150,12 +150,12 @@
         <default key="deleted">0</default>
     </route>
     <route id="ems_content_revisions_view" path="/revisions/{type}:{ouuid}/{revisionId}/{compareId}" methods="GET">
-        <default key="_controller">EMS\CoreBundle\Controller\ContentManagement\DataController:revisionsDataAction</default>
+        <default key="_controller">EMS\CoreBundle\Controller\Revision\DetailController:detailRevision</default>
         <default key="revisionId">0</default>
         <default key="compareId">0</default>
     </route>
     <route id="data.revisions" path="/revisions/{type}:{ouuid}/{revisionId}/{compareId}" methods="GET">
-        <default key="_controller">EMS\CoreBundle\Controller\ContentManagement\DataController:revisionsDataAction</default>
+        <default key="_controller">EMS\CoreBundle\Controller\Revision\DetailController:detailRevision</default>
         <default key="revisionId">0</default>
         <default key="compareId">0</default>
     </route>
